style[OK-95] Move UndoRedoCommand interface to another file
