+++ client/src/app/globals.css
@@ -136,7 +136,6 @@ input[type="color"]::-webkit-color-swatch-wrapper {
   animation: slideUp 0.5s ease-in forwards;
 }
 
-
 .zoom-badge {
   position: absolute;
   bottom: 1em;
@@ -144,11 +143,13 @@ input[type="color"]::-webkit-color-swatch-wrapper {
   background-color: rgba(55, 65, 81, 0.9);
   color: white;
   padding: 0.5rem 1rem;
-  width:10em;
+  width: 10em;
   border-radius: 0.5rem;
   font-size: 1.125rem;
-  transition: opacity 0.5s ease, filter 0.5s ease;
-  pointer-events: none
+  transition:
+    opacity 0.5s ease,
+    filter 0.5s ease;
+  pointer-events: none;
 }
 
 .zoom-badge.show {
