--- client/src/interfaces/canvas-action.ts
@@ -1,4 +1,0 @@
-export interface UndoRedoCommand {
-  redo: () => void;
-  undo: () => void;
-}
