+++ client/src/interfaces/undo-redo-context.ts
@@ -1,9 +1,10 @@
-import { UndoRedoCommand } from "./undo-redo-command";
-
 export interface UndoRedoContext {
   saveCommand: (command: UndoRedoCommand) => void;
   undo: () => void;
   redo: () => void;
-  listenersOn: boolean;
-  setListenersOn: (newValue: boolean) => void;
+}
+
+export interface UndoRedoCommand {
+  redo: () => void;
+  undo: () => void;
 }
